major refactor
This commit includes a major refactors and patches. The markup lexer is approaching a very stable version. Changes are as follows:

- Sugar utils for querying parsed types/tokens
- Purged the remaining template languages
- 1/2 Liquid comments refactor. This essentially aligns Liquid comments with HTML comments and also allows us to pass in inline rules.
- Solved the vast majority of edge cases pertaining to Liquid code contained within HTML attributes. The logic is much cleaner, more precise and flexible.

The main updates and patches here are solely related to attributes formatting. Wherein Prettify will now gracefully indent Liquid infused within HTML Attribute values. It does not matter the _type_ of liquid code being applied as they are treated as regular HTML attributes but with some slight differences and far more control.

This commit and lexer patch also ships inline formatting support using `@prettify format: ....` and while on the nexus of comments, all previous issues relating to Liquid comments are practically solved/fixed. There is still some work to be done but the project is almost ready for the big time.