fixes a couple of defects for attributeChain
In addition, this commit also addresses the PrettyDiff issue caused by Angle Brackets being processed as HTML delimiters and thus breaking the parse when lexing Liquid attributes.

The actual issue was submitted for Twig but also wreaked havoc in Liquid. This defect was fixed by keeping track of entry points of tokens, ie: when we are within a Liquid template tag ({{ or {%) > (%} or }}) then we assigned a variable which allows us to skip interception of angle brackets that were being used in control based conditions like {% if x > 1 %}

See:
https://github.com/prettydiff/prettydiff/issues/621